1.视频播放支持缩放和拖拽 2.opengl矩阵为列优先存储机制

+++ app/src/main/java/com/norman/android/hdrsample/player/decode/base/FFmpegDecode.kt
@@ -181,7 +181,7 @@ class FFmpegDecode(private val mimeType: String, private val fileSource: FileSou
         return null
     }
 
-    override fun onPlayProgress(frame: ByteBuffer?, time: Double) {
+    override fun onPlayProgress(frame: ByteBuffer?, time: Long) {
         runBlocking {
             suspendCancellableCoroutine<Boolean> { continuation ->
                 createLooper?.also {
